Renaming Components to right names
