+++ fantom-homes_react/src/Components/HomesCard/HomesCard.css
