+++ fantom-homes_react/src/Components/HomesCard/data.js
